server/mfa_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
webauthn==2.8.0a1
cryptography>=44.0.2
bcrypt==4.1.1
pyjwt==2.11.0
python-jose==3.3.0
python-multipart==0.0.6
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0
